tests/wpt/web-platform-tests/css/css-grid/alignment/grid-self-alignment-non-static-positioned-items-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: Self-Alignment of non-static absolute positioned items</title>
<link rel="author" title="Javier Fernandez Garcia-Boente" href="mailto:jfernandez@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#column-align">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#abspos-items">
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-justify-self">
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-align-self">
<meta name="assert" content="Alignment properties don't apply on absolute positioned grid items with non-static position in the property's axis.">
<style>
.grid {
  position: relative;
  display: inline-grid;
  grid-template-columns: 100px 150px;
  grid-template-rows: 150px 100px;
  font: 10px/1 Ahem;
  background: grey;
}
.grid > div { position: absolute; }
.grid > :nth-child(1) { background: green; }
.grid > :nth-child(2) { background: blue; }
.grid > :nth-child(3) { background: yellow; }
.grid > :nth-child(4) { background: red; }

.RTL { direction: rtl; }
.verticalLR { writing-mode: vertical-lr; }
.verticalRL { writing-mode: vertical-rl; }

.firstRowFirstColumn {
  grid-row: 1 / 2;
  grid-column: 1 / 2;
  align-self: start;
  justify-self: right;
  left: 5px;
}
.firstRowSecondColumn {
  grid-row: 1 / 2;
  grid-column: 2 / 3;
  align-self: end;
  justify-self: center;
  right: 10px;
}
.secondRowFirstColumn {
  grid-row: 2 / 3;
  grid-column: 1 / 2;
  align-self: center;
  justify-self: start;
  top: 10px
}
.secondRowSecondColumn {
  grid-row: 2 / 3;
  grid-column: 2 / 3;
  align-self: end;
  justify-self: left;
  left: 5px;
  top: 10px;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('.grid')">
<div class="grid">
  <div data-offset-x="5"   data-offset-y="0"   data-expected-width="60" data-expected-height="10" class="firstRowFirstColumn">X XX X</div>
  <div data-offset-x="170" data-offset-y="120" data-expected-width="70" data-expected-height="30" class="firstRowSecondColumn">XX X<br>X XXX X<br>XX XXX</div>
  <div data-offset-x="0"   data-offset-y="160" data-expected-width="60" data-expected-height="10" class="secondRowFirstColumn">X XX X</div>
  <div data-offset-x="105" data-offset-y="160" data-expected-width="60" data-expected-height="40" class="secondRowSecondColumn">XX X<br>X XXX<br>X<br>XX XXX</div>
</div>

<div class="grid RTL">
  <div data-offset-x="155" data-offset-y="0"   data-expected-width="60" data-expected-height="10" class="firstRowFirstColumn">X XX X</div>
  <div data-offset-x="70"  data-offset-y="120" data-expected-width="70" data-expected-height="30" class="firstRowSecondColumn">XX X<br>X XXX X<br>XX XXX</div>
  <div data-offset-x="190" data-offset-y="160" data-expected-width="60" data-expected-height="10" class="secondRowFirstColumn">X XX X</div>
  <div data-offset-x="5"   data-offset-y="160" data-expected-width="60" data-expected-height="40" class="secondRowSecondColumn">XX X<br>X XXX<br>X<br>XX XXX</div>
</div>

<br><br>

<div class="grid verticalLR">
  <div data-offset-x="5"   data-offset-y="40"  data-expected-width="10" data-expected-height="60" class="firstRowFirstColumn">X XX X</div>
  <div data-offset-x="110" data-offset-y="140" data-expected-width="30" data-expected-height="70" class="firstRowSecondColumn">XX X<br>X XXX X<br>XX XXX</div>
  <div data-offset-x="195" data-offset-y="10"  data-expected-width="10" data-expected-height="60" class="secondRowFirstColumn">X XX X</div>
  <div data-offset-x="155" data-offset-y="110" data-expected-width="40" data-expected-height="60" class="secondRowSecondColumn">XX X<br>X XXX<br>X<br>XX XXX</div>
</div>

<div class="grid verticalRL">
  <div data-offset-x="105" data-offset-y="40"  data-expected-width="10" data-expected-height="60" class="firstRowFirstColumn">X XX X</div>
  <div data-offset-x="210" data-offset-y="140" data-expected-width="30" data-expected-height="70" class="firstRowSecondColumn">XX X<br>X XXX X<br>XX XXX</div>
  <div data-offset-x="45"  data-offset-y="10"  data-expected-width="10" data-expected-height="60" class="secondRowFirstColumn">X XX X</div>
  <div data-offset-x="5"   data-offset-y="110" data-expected-width="40" data-expected-height="60" class="secondRowSecondColumn">XX X<br>X XXX<br>X<br>XX XXX</div>
</div>
</body>
